+++ packages/content/src/index.ts
@@ -1,4 +1,2 @@
 export * from './Content';
 export * from './ContentBleed';
-export * from './PageContent';
-export * from './SectionContent';
